Break AbstractDataReader into two classes: AbstractColumnReader and AbstractDataReader.
